+++ server/backend/app/models/group_user_association.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Table, Column, Integer, ForeignKey, Boolean, String
+from sqlalchemy import Table, Column, Integer, ForeignKey, BigInteger
 from db.base import Base
 
 association_table = Table(
     'group_user_association',
     Base.metadata,
     Column('group_id', Integer, ForeignKey('groups.id')),
-    Column('user_id', Integer, ForeignKey('users.id')),
+    Column('user_id', BigInteger, ForeignKey('users.id')),
     Column('hierarchy_level', Integer)  
 )
